move HTTPNetworkingRedirectionError to its own file + tests
